+++ packages/ente-api/src/templates/InvitationLink.de.ts
@@ -1,4 +1,6 @@
 import * as Handlebars from "handlebars";
+import "./register-helpers";
+import { Roles } from "ente-types";
 
 const title = "Sie wurden zu EntE eingeladen.";
 
@@ -15,8 +17,26 @@ const template = Handlebars.compile(`
         </mj-text>
         
         <mj-text>
+          {{#ifEqual role "${Roles.PARENT}"}}
           Sie wurden eingeladen, auf EntE die Fehlstunden ihrer Kinder zu verwalten.
           Um EntE benutzen zu können, setzen Sie bitte ihr Passwort mithilfe des folgenden Links.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.STUDENT}"}}
+          Sie wurden eingeladen, auf EntE ihre Fehlstunden zu verwalten.
+          Um EntE benutzen zu können, setzen Sie bitte ihr Passwort mithilfe des folgenden Links.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.MANAGER}"}}
+          Sie wurden eingeladen, auf EntE die Fehlstunden ihrer Stufe zu verwalten.
+          Um EntE benutzen zu können, setzen Sie bitte ihr Passwort mithilfe des folgenden Links.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.TEACHER}"}}
+          Sie wurden eingeladen, auf EntE die Fehlstunden ihrer Schüler einzusehen.
+          Um EntE benutzen zu können, setzen Sie bitte ihr Passwort mithilfe des folgenden Links.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.ADMIN}"}}
+          Sie wurden eingeladen, eine EntE-Instanz zu administrieren.
+          Um EntE benutzen zu können, setzen Sie bitte ihr Passwort mithilfe des folgenden Links.
+          {{/ifEqual}}
           
           <br /> <br />