+++ packages/ente-api/src/templates/InvitationLink.en.ts
@@ -1,4 +1,6 @@
 import * as Handlebars from "handlebars";
+import "./register-helpers";
+import { Roles } from "ente-types";
 
 const title = "You have been invited to EntE.";
 
@@ -15,8 +17,26 @@ const template = Handlebars.compile(`
         </mj-text>
         
         <mj-text>
+          {{#ifEqual role "${Roles.PARENT}"}}
           You have been invited to manage your children's missed lessons on EntE.
           In order to be able to use EntE, please set your password using the following link.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.STUDENT}"}}
+          You have been invited to manage your missed lessons on EntE.
+          In order to be able to use EntE, please set your password using the following link.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.MANAGER}"}}
+          You have been invited to manage your students missed lessons on EntE.
+          In order to be able to use EntE, please set your password using the following link.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.TEACHER}"}}
+          You have been invited to view your students missed lessons on EntE.
+          In order to be able to use EntE, please set your password using the following link.
+          {{/ifEqual}}
+          {{#ifEqual role "${Roles.ADMIN}"}}
+          You have been invited to administer an EntE-Instance.
+          In order to be able to use EntE, please set your password using the following link.
+          {{/ifEqual}}
           
           <br /> <br />