+++ packages/ente-api/src/templates/InvitationLink.ts
@@ -7,7 +7,7 @@
  */
 
 // tslint:disable-next-line:no-var-requires
-import { Languages, getByLanguage } from "ente-types";
+import { Languages, getByLanguage, Roles } from "ente-types";
 import { InvitationLinkEN } from "./InvitationLink.en";
 import { InvitationLinkDE } from "./InvitationLink.de";
 import { mjml2html } from "../helpers/mjml";
@@ -17,9 +17,17 @@ const getTemplate = getByLanguage({
   [Languages.GERMAN]: InvitationLinkDE
 });
 
-export const InvitationLink = (linkAddress: string, lang: Languages) => {
+export const InvitationLink = (
+  linkAddress: string,
+  userRole: Roles,
+  lang: Languages
+) => {
   const { template, title } = getTemplate(lang);
-  const mjml = template({ linkAddress, linkDisplay: linkAddress });
+  const mjml = template({
+    linkAddress,
+    linkDisplay: linkAddress,
+    role: userRole
+  });
   const { errors, html } = mjml2html(mjml);
   if (errors.length > 0) throw new Error("MJML Error");
   return { html, subject: title };