+++ packages/ente-api/src/templates/register-helpers.ts
@@ -1,0 +1,5 @@
+import * as Handlebars from "handlebars";
+
+Handlebars.registerHelper("ifEqual", function(this: any, arg1, arg2, options) {
+  return arg1 === arg2 ? options.fn(this) : options.inverse(this);
+});